improvement(data-drains): extract normalizePrefix and buildObjectKey to shared utils
